cleanup: separate some util.h functions
also avoid include useless functions when possible (faster compiles)
